Add comprehensive tests for GeminiCliAgent and OpenCodeAgent functionality
This commit introduces a series of tests for the GeminiCliAgent and OpenCodeAgent, enhancing the coverage of hook matching, session management, and command execution. New tests validate the behavior of hook helpers, transcript handling, and error scenarios for missing or mismatched items in knowledge references. Additionally, the OpenCodeAgent's session path and metadata handling are thoroughly tested, ensuring robust functionality. These changes improve the reliability and maintainability of the agents by ensuring they perform as expected under various conditions.
