+++ bitloops/src/adapters/agents/open_code.rs
@@ -1,6 +1,11 @@
-pub mod agent;
+pub mod agent_api;
+mod agent_hook_support;
+mod agent_runtime;
+mod agent_session_io;
 pub mod cli_commands;
 pub mod hooks;
 pub mod plugin;
 pub mod transcript;
 pub mod types;
+
+pub use agent_api as agent;
